+++ .gitignore
@@ -211,3 +211,5 @@ Temporary Items
 
 # Ignore Gradle build output directory
 build
+
+src/main/resources/application.properties
