+++ build.gradle
@@ -17,8 +17,13 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-web'
 	developmentOnly 'org.springframework.boot:spring-boot-devtools'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
+	implementation('org.springframework.boot:spring-boot-starter-data-jpa')
+	runtimeOnly('org.postgresql:postgresql')
 }
 
 tasks.named('test') {
 	useJUnitPlatform()
 }
+
+
+
