--- src/main/java/com/rpapSongr/helloWorld/Album.java
@@ -1,82 +1,0 @@
-package com.rpapSongr.helloWorld;
-
-import java.awt.*;
-import java.net.URL;
-
-public class Album {
-    private String album;
-    private String title;
-    private String artist;
-    private Integer songCount;
-    private Double length;
-    private String imageUrl;
-
-    public Album(String album, String title, String artist, Integer songCount, Double length, String imageUrl) {
-        this.album = album;
-        this.title = title;
-        this.artist = artist;
-        this.songCount = songCount;
-        this.length = length;
-        this.imageUrl = imageUrl;
-    }
-
-    public String getAlbum() {
-        return album;
-    }
-
-    public void setAlbum(String album) {
-        this.album = album;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getArtist() {
-        return artist;
-    }
-
-    public void setArtist(String artist) {
-        this.artist = artist;
-    }
-
-    public Integer getSongCount() {
-        return songCount;
-    }
-
-    public void setSongCount(Integer songCount) {
-        this.songCount = songCount;
-    }
-
-    public Double getLength() {
-        return length;
-    }
-
-    public void setLength(Double length) {
-        this.length = length;
-    }
-
-    public String getImageUrl() {
-        return imageUrl;
-    }
-
-    public void setImageUrl(String imageUrl) {
-        this.imageUrl = imageUrl;
-    }
-
-    @Override
-    public String toString() {
-        return "Album{" +
-                "album='" + album + '\'' +
-                ", title='" + title + '\'' +
-                ", artist='" + artist + '\'' +
-                ", songCount=" + songCount +
-                ", length=" + length +
-                ", imageUrl='" + imageUrl + '\'' +
-                '}';
-    }
-}
