+++ src/main/resources/application.properties
@@ -1,1 +1,5 @@
 
+spring.datasource.url=jdbc:postgresql://localhost:5432/albums1
+
+spring.jpa.hibernate.dll-auto=update
+spring.jpa.generate-ddl=true
