+++ src/main/resources/templates/welcome.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <title>Title</title>
+  <link href="/css/style.css" rel="stylesheet">
+</head>
+<body>
+<h1>Welcome to Songr!</h1>
+</body>
+</html>
